api fix for int conversion and publish related changes

+++ BitesByte_API/Model/MaintainanceCaloriesDTO.cs
@@ -3,7 +3,7 @@
     public class MaintainanceCaloriesDTO
     {
         public decimal weight {  get; set; }
-        public decimal height { get; set; }
+        public int height { get; set; }
         public int age { get; set; }
         public string gender { get; set; }
         public string activityLevel { get; set; }
